Organize churn and retention into an engagement module
This reorganize churn and retention into a shared module since their
lineage is closely linked.

Tests have also been reformatted to include a shared `data.py` file that
contains the default sample and schema for `main_summary` and
`new_profile`. The idiom of `test_transform` is reused in daily
retention.

Post-Review: Fix code consistency issues

These issues are mostly related to PEP8.
* Avoid star imports
* Use a more style of imports compatible with Python 3
* Be more consistent with quote usage in `engagement.retention.job`

In general, global constant variables are imported into local scope,
while functions are used in the original namespace across the
changed files.
