+++ mozetl/cli.py
@@ -1,11 +1,12 @@
 import click
-from mozetl.churn import churn
-from mozetl.search import dashboard, search_rollups
-from mozetl.taar import taar_locale, taar_similarity
-from mozetl.sync import bookmark_validation
+
 from mozetl.clientsdaily import rollup as clientsdaily
+from mozetl.engagement.churn import job as churn_job
+from mozetl.engagement.retention import job as retention_job
 from mozetl.experimentsdaily import rollup as experimentsdaily
-from mozetl.retention import daily_retention
+from mozetl.search import dashboard, search_rollups
+from mozetl.sync import bookmark_validation
+from mozetl.taar import taar_locale, taar_similarity
 
 
 @click.group()
@@ -13,10 +14,10 @@ def entry_point():
     pass
 
 
-entry_point.add_command(churn.main, "churn")
+entry_point.add_command(churn_job.main, "churn")
 entry_point.add_command(clientsdaily.main, "clients_daily")
 entry_point.add_command(experimentsdaily.main, "experiments_daily")
-entry_point.add_command(daily_retention.main, "retention")
+entry_point.add_command(retention_job.main, "retention")
 entry_point.add_command(dashboard.main, "search_dashboard")
 entry_point.add_command(search_rollups.main, "search_rollup")
 entry_point.add_command(bookmark_validation.main, "sync_bookmark_validation")
